Refactor TypeSafeFacebookToken to extract generic TypeSafeFacebook elements
